Improved error handling on rover server
When starting the rover roboticsnet server, the new system attempts to initialize each critical object (umanager, webcam stream object) in separate try blocks. If an exception is raised, the server will start without passing in the hooks associated with that component. This way, the server can be run without a motor arduino connected or even a webcam theoretically.